hitch::utils

Module authorization

Functions

get_current_user

Get the current user’s email from git configuration

get_remaining_approvers

Get users who can still approve the request

is_authorized_approver

Check if a user is authorized to approve requests for an environment

validate_approval_authorization

Validate that a user can approve a specific request

validate_cancellation_authorization

Validate that a user can cancel a request

validate_rejection_authorization

Validate that a user can reject a request